Make SSR a rendering pipeline
