+++ packages/dev/core/src/PostProcesses/RenderPipeline/Pipelines/index.ts
@@ -3,3 +3,4 @@ export * from "./lensRenderingPipeline";
 export * from "./ssao2RenderingPipeline";
 export * from "./ssaoRenderingPipeline";
 export * from "./standardRenderingPipeline";
+export * from "./ssrRenderingPipeline";
