+++ packages/dev/core/src/PostProcesses/index.ts
@@ -31,5 +31,4 @@ export * from "./volumetricLightScatteringPostProcess";
 export * from "./vrDistortionCorrectionPostProcess";
 export * from "./vrMultiviewToSingleviewPostProcess";
 export * from "./screenSpaceReflectionPostProcess";
-export * from "./screenSpaceReflection2PostProcess";
 export * from "./screenSpaceCurvaturePostProcess";
